% Generated by roxygen2 (4.0.0): do not edit by hand
\name{DeltaBounds}
\alias{DeltaBounds}
\title{Calculates matching-based bounds on the natural direct effects.}
\usage{
DeltaBounds(Z, Y, S, W, S.cat = NULL, Gamma = 1, Psi = 1,
  NDE = "Delta1", R = 500, conf = 0.95)
}
\arguments{
\item{Z}{A vector of treatment indicators.}

\item{Y}{A vector of (binary) outcomes.}

\item{S}{A vector of biomarker values, i.e., potential mediators}

\item{W}{A vector or matrix of covariates to be used in the matching}

\item{S.cat}{(Optional) vector containing a coarsened version of \code{S}. There should be at least one treated and one untreated subject within each level of \code{S.cat}.
By default (\code{S.cat=NULL}), \code{S} is split into quartiles.}

\item{Gamma}{The sensitivity parameter controlling the odds ratio between the true and estimated odds of a discordant pair favoring the treatment. Default is 1 (corresponding to no unmeasured confonding), values >= 1 are valid.}

\item{Psi}{The sensitivity parameter controlling the odds ratio between the true and estimated odds of a pair being discordant. Default is 1 (corresponding to no unmeasured confonding), values >= 1 are valid.}

\item{NDE}{Specifies which natural direct effect (NDE) to calculate. Valid options are \code{"Delta1"} (the default)
and \code{"Delta0"}, corresponding to \eqn{E(Y^{1S(1)} - Y^{0S(1)})} and \eqn{E(Y^{1S(0)} - Y^{0S(0)})} respectively.}

\item{R}{Number of bootstrap samples used to estimate uncertainty intervals. Defaults to \code{500}.}

\item{conf}{Confidence level for the bootstrapped uncertainty intervals. Defaults to \code{0.95}.}
}
\value{
A vector of five elements containing:
 \describe{
 \item{\code{delta}}{The point estimate for the NDE.}
 \item{\code{delta.II}}{The ignorance interval for the NDE, reflecting only uncertainty due to unmeasured confounding as reflected by \code{Gamma} and \code{Psi}.}
 \item{\code{delta.CI}}{The confidence interval for the NDE, reflecting only stochastic uncertainty, ignoring uncertainty due to unmeasured confounding.}
 \item{\code{delta.UI}}{The uncertainty interval for the NDE, reflecting both sources of uncertainty.}
}
}
\description{
This function estimates matching-based bounds on the two natural direct effects
}
\details{
This function computes uncetainty intervals for natural direct effects which incorporate uncertainty in the point estimate
due to both possible unmeasured confounding and statistical uncertainty in estimation due to the matching procedure.
Currently, percentile-based bootstrap uncertainty intervals are returned as they appear to yield somewhat better coverage than
alternatives.
}

