% Generated by roxygen2 (4.0.0): do not edit by hand
\name{propScores}
\alias{propScores}
\title{Calculate propensity score for matching.}
\usage{
propScores(dat)
}
\arguments{
\item{dat}{A data frame containing columns for the treatment (\code{Z}), outcome (\code{Y}),
potential mediator (\code{S} and \code{S.cat}), and baseline covariates (\code{W1, W2, ...})}
}
\value{
A vector of propensity scores of the same length as \code{Z}.
}
\description{
Computes the propensity score for treatment based on covariates. Should not generally be called directly.
}
\details{
The propensity scores are computed by fitting a logistic regression model to the binary outcome \code{Y},
 with linear terms for each of the baseline covariates \code{W1, W2, ...}
}

